requirements.txt
streamlit
python-dotenv
azure-ai-inference
